wip: make ipc_service sample functional with flpr
At the moment the sample starts and is logging with STM from App and
FLPR. However, FLPR faults with an inaccessible instruction after
random time. The registers' content analysis suggests its around
sys_clock_cycle_get_32(), after the value is read from GRTC and copied
to a0. However, this function is frequently called in a busy wait in the
main thread, so it might be just a coincidence that the fault always
hits the most frequently executed operation.

Idea for further debugging:
* use longer intervals to avoid busy waiting in the main thread
* use GPIO pins to switch on/off leds around suspected code sequences
and see which instruction results in the fault (LED stays enabled)